fix(onboarding): address PR feedback
- onboarding-shell: replace Button asChild (not supported) with the
  Base UI render={<a/>} pattern used elsewhere in the codebase. Fixes
  the typecheck + build failures.
- handler/agents_create.go: extract subagent UUID parsing + attach
  logic into agents_create_subagents.go so the file lands at 272
  lines (was 305, over the 300-line cap).
- handler/agents_create_test.go: move the six is_employee /
  subagent_ids tests + the seedAgent helper into
  agents_create_subagents_test.go so the file lands at 186 lines (was
  347).
- docs/docs.go: revert to upstream — auto-generated, should be
  excluded from the file-length check, not regenerated in this PR.
