+++ internal/handler/agents_create.go
@@ -177,22 +177,10 @@ func (h *AgentHandler) Create(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var subagentUUIDs []uuid.UUID
-	if len(req.SubagentIDs) > 0 {
-		subagentUUIDs = make([]uuid.UUID, 0, len(req.SubagentIDs))
-		seen := make(map[uuid.UUID]struct{}, len(req.SubagentIDs))
-		for _, raw := range req.SubagentIDs {
-			parsed, parseErr := uuid.Parse(raw)
-			if parseErr != nil {
-				writeJSON(w, http.StatusBadRequest, map[string]string{"error": fmt.Sprintf("invalid subagent_id %q", raw)})
-				return
-			}
-			if _, dup := seen[parsed]; dup {
-				continue
-			}
-			seen[parsed] = struct{}{}
-			subagentUUIDs = append(subagentUUIDs, parsed)
-		}
+	subagentUUIDs, subErrMsg := parseSubagentIDs(req.SubagentIDs)
+	if subErrMsg != "" {
+		writeJSON(w, http.StatusBadRequest, map[string]string{"error": subErrMsg})
+		return
 	}
 
 	var skillUUIDs []uuid.UUID
@@ -244,29 +232,8 @@ func (h *AgentHandler) Create(w http.ResponseWriter, r *http.Request) {
 			}
 		}
 
-		if len(subagentUUIDs) > 0 {
-			var visibleSubagents []model.Agent
-			if err := tx.
-				Select("id", "is_employee").
-				Where("id IN ? AND org_id = ? AND status = ?", subagentUUIDs, org.ID, "active").
-				Find(&visibleSubagents).Error; err != nil {
-				return fmt.Errorf("validate subagent_ids: %w", err)
-			}
-			if len(visibleSubagents) != len(subagentUUIDs) {
-				return fmt.Errorf("one or more subagent_ids are not active agents in this workspace")
-			}
-			for _, sub := range visibleSubagents {
-				if sub.IsEmployee {
-					return fmt.Errorf("subagent_id %s refers to an employee; employees cannot be subagents", sub.ID)
-				}
-			}
-			links := make([]model.AgentSubagent, len(visibleSubagents))
-			for i, sub := range visibleSubagents {
-				links[i] = model.AgentSubagent{AgentID: agent.ID, SubagentID: sub.ID}
-			}
-			if err := tx.Create(&links).Error; err != nil {
-				return fmt.Errorf("attach subagents: %w", err)
-			}
+		if err := attachSubagents(tx, org.ID, agent.ID, subagentUUIDs); err != nil {
+			return err
 		}
 
 		if err := createAgentTriggers(tx, org.ID, agent.ID, req.Triggers); err != nil {
